+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - unreleased
+
+### Changed
+
+- Use hand-drawn label images
+
+## [0.6.0] - 2025-11-28
+
+### Added
+
+- Prepare 24 doors
+- Prepare first 6 content packages
+
 ## [0.5.0] - 2025-11-27
 
 ### Fixed